refactor(view-engine): move viewServiceContract to contracts layer
