+++ packages/view-engine/src/record/IndexedDBViewHost.ts
@@ -16,7 +16,7 @@ import {
   type StatefulViewHostOptions,
   type ViewStateTransaction,
 } from './StatefulViewHost.js';
-import { ViewServiceError } from './viewServiceContract.js';
+import { ViewServiceError } from '../contracts/viewServiceContract.js';
 import { message } from '../lib/snapshot.js';
 export interface IndexedDBViewHostOptions extends StatefulViewHostOptions {
   databaseName?: string;
