+++ packages/view-engine/test/viewPage.management.test.tsx
@@ -22,7 +22,7 @@ import {
 } from '@testing-library/react';
 import { afterEach, expect, it, vi } from 'vitest';
 import { ViewPage } from './fixtures/OwnedViewPage.js';
-import { ViewServiceError } from '../src/record/viewServiceContract.js';
+import { ViewServiceError } from '../src/contracts/viewServiceContract.js';
 import { instance, setup } from './fixtures/viewPage.js';
 
 afterEach(cleanup);
