+++ packages/view-engine/test/viewPage.pendingCreates.test.tsx
@@ -21,7 +21,7 @@ import {
 } from '@testing-library/react';
 import { afterEach, expect, it, vi } from 'vitest';
 import { ViewPageContent } from '../src/view/ViewPageContent.js';
-import { ViewServiceError } from '../src/record/viewServiceContract.js';
+import { ViewServiceError } from '../src/contracts/viewServiceContract.js';
 import { instance, setup } from './engine/fixtures.js';
 
 afterEach(cleanup);
